requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.28
typing-extensions==4.0.1
Werkzeug==2.0.2
